Javascript Revamp
instead of exporting functions now we'll export objects. These objects will essentially be view/component files. Each file will render its own view and attach all listeners/functionality.

This was done so we can bundle all the code in one object and export it in one line, as opposed to each function seperately. Also it helps that we can render in-place as opposed to before
