+++ src/scripts/menu.js
@@ -1,14 +1,16 @@
+//this page should contain all listeners and functionality for the main menu page
+
 export function menu() {
     // const nudityToggle = document.getElementById("nudityToggleSwitch");
-    // const profanityToggle = document.getElementById("profanityToggleSwitch");
+    // const soundToggle = document.getElementById("soundToggleSwitch");
 
-    // if (nudityToggle && profanityToggle) {
+    // if (nudityToggle && soundToggle) {
     //     nudityToggle.addEventListener("click", () => {
     //         alert("nudity");
     //     });
 
-    //     profanityToggle.addEventListener("click", () => {
-    //         alert("profanity");
+    //     soundToggle.addEventListener("click", () => {
+    //         alert("sound");
     //     });
     // } else {
     //     alert("Error, App wasn't able to load properly, please re-install");
