+++ src/scripts/settings.js
@@ -1,0 +1,4 @@
+export function settingsView() {
+    //add macro listener
+
+}
