+++ src/templates/menu.html
@@ -1,13 +1,14 @@
 <div class="menu-options">
   <div class="option-row">
-    <label>Nudity Filter: </label>
+    <label>Video Filter: </label>
     <input type="checkbox"></input>
   </div>
   <div class="option-row">
-    <label>Profanity Filter: </label>
+    <label>Sound Filter: </label>
     <input type="checkbox"></input>
   </div>
 </div>
+
 <!-- 
 <div class="user-info">
   <p>Jane Doe</p>
