+++ src/templates/navbar.html
@@ -1,4 +1,4 @@
 <nav class="navbar">
     <h1>Project Censor</h1>
-    <a href="#" id="settingsButton"><i class="fa-solid fa-gear"></i></a>
+    <button class="icon-button" id="settingsButton"><i class="fa-solid fa-gear"></i></button>
 </nav>
