+++ src/templates/settings.html
@@ -1,0 +1,19 @@
+<div class="menu-options">
+    <div class="option-row">
+      <label>Video Filter Strength: </label>
+      <select>
+        <option>-- Medium (Default) --</option>
+      </select>
+    </div>
+    <div class="option-row">
+        <label>Sound Filter Strength: </label>
+        <select>
+          <option>-- Medium (Default) --</option>
+        </select>
+    </div>
+    <div class="option-row">
+        <label>Video Filter Macro: </label>
+        <div class="macro-input"></div>
+        <button class="icon-button" id="btnEditVideoMacro"><i class="fa-solid fa-pen"></i></button>
+    </div>
+</div>
